Ran structured ice shelf with no thermodynamics.
I turned off the neumann bc for temperature and salinity
along the top boundary. The model didnt reach the first time step.
Immediately the residual dropped sufficiently low (1e-12) so that
the next residual was the same order of magnitude. This meant
the relative error stopping criterion was not met.
 I believe this means the model is already at a steady
state.
The error message was:
-----------------------------------------------------------------
firedrake.exceptions.ConvergenceError: Nonlinear solve failed to
converge after 25 nonlinear iterations.
Reason:
   DIVERGED_LINE_SEARCH
-----------------------------------------------------------------

Nb I made a mistake in the last commit message. I had been running
with an eddy viscosity and diffusivity of 1e-2m^2/s not 10e-2m^2/s.

Also most of the comments in ice_shelf_cavity_structured_z.py are
misleading or wrong. Needs to be fixed!