Ran structured ice shelf with restoring imbalance and no thermodynamics
Ran structured ice shelf with initial constant Salinity conditions of
34.4 PSU and restoring conditions of 34.5 PSU. Expected to get clockwise
overturning because of pressure difference generated at bottom on rhs.
The initial conditions and restoring values for temperature were equal
at 1degC.

Ran with dt = 60s K = 1E-2m^2/s and outputed every 1200s.

The solution had already blown up in a similar way to before when melting
was turned on. i.e stripes in density and velocity components. Presumably
the pattern emerges because the grid supports an instability like this, but
it might not be the root cause.

The blow up definitely seems to be concentrated at the RHS open boundary to begin with
and move towards the grounding line. At 3600 seconds there were no stripes in density
or velocity at the grounding line.

To do:
-  Run and look at the first ten time steps to see if it possible to work out what is
going on...
- Run with lower restoring salinity of 34.3 PSU. is initial flow reversed?
- Run equivalent just with closed box. What happens?
- Run equivalent with restoring turned off. Just with bc. What happens?
- Run scalar advection example. does this still run?
- What happens if I turn diffusion off?
- What happens if I turn advection off?

To think about:
- Does this demonstrate that melting/influence of the top boundary is not the problem?
- Does this demonstrate that the problem is with the open boundary
since the blow up definitely happens there first?
- Is it the restoring region or the bc?