Is scalar field the problem?
Ran structured ice shelf with restoring imbalance and no thermodynamics
again but looked at first ten time steps. It seemed to me that the
velocity field was behaving fine on the first time step with clockwise
flow. The flow in/out was spread over the whole column and not concentrated
at the top and bottom which seemed to be the case once the solution had
blown up.

The density however already had one set of stripes on the first time step.
Does this mean that there is a problem in the density code somewhere?

Possibly running with lower diffusivities (1E-2 is the lowest they have been)
exposes this problem...

To do
- Run same ten time steps but with density coupling off. i.e set
\pho' = 0. Need to also set the stress open bc to zero.

Hopefully the density field will still have stripes which will be advected
by a velocity field that isn't affected by the density.