Scalar field probably not the problem...
Ran structured ice shelf with no thermodynamics and with
mom_source vector as (0,0). Initially also ran with stress
boundary = -n*0. But this was immediately at a steady state since
I wasn't provide a forcing! The density inside the box and outside
were the same!

I added the original stress boundary back in and ran. This is equivalent
to setting the density inside to the standard density 1030 kg/m^3 perturbed
to the reference values of salinity and temperature i.e T = 0degC and S = 35PSU.

Therefore would expect anticlockwise flow when the outside restoring region is set
to T = 1degC and S = 34.5PSU because both of these lead to reducing the density outside
the domain.

The flow was anticlockwise and initially quite fast u_mag = 0.5m/s. There were also stripes
developing after the first time step up to the 10th in both the density and velocity fields.
The orientation of the stripes was changed though so that they slanted like \, which is
presumably because of the anticlockwise flow.

It possibly answers the question that the density field isn't to blame because the velocity
still blows up. Having said this I don't know if the initialisation shock might have been too large.

The horizontal grid Re = U dx / nu = 5e-1 * 5e1 / 1e-2  = 2500 !!
vertical grid Re = W dz / nu = 1e-1 * 5e-1 / 1e-2 = 5

horizontal cfl = Udt /dx = 5e-1 * 6e1 / 5e1 = 0.6
vertical cfl = W dt / dx = 1e-1 * 6e1 / 5e-1 = 12 !!

so really I should expect the simulation to blow up.

Therefore it is probably worth trying with the initial mom_source set to the proper initial
conditions so that the forcing is only due to the 0.1 PSU difference. Hopefully the flow
speeds will be slower and the velocity solve will be stable...

Then move onto closing the box and turning off restoring (separately!) if the velocity
blows up again.