Strong bcs no slip rhs makes no difference...
Added strong bc back in to enforce no slip along the rhs wall. This is
because the ocean would resist vertical shearing.

It doesn't seem to have made any difference though. I was hoping this
would reduce the vertical velocity. I think I implemented
it correctly. May come back to this later but will move on for now to
reducing delPSU. I will leave the strong bc in.