Run with del PSU = 0.0001 Redx = 0.3 still stripy...
So ran with del PSU = 0.0001 and again the velocities were an order
of magnitude lower, so the Redx = 0.3.

The stripes still occur in the velocity and density fields.

This means they are not caused by the grid Reynolds number being too high.
This is backed up by the full melting simulation failing with stripes even
 though the velocities were order 1e-7m/s.

Is the low viscosity/diffusivity exposing a bug?

If it is with the open boundary then what happens if I close the box? Before
closing the box with the melt simulation didnt change much. Now it is a simpler
setup so it would be worth trying to run this with a closed box with delPSU =
0.0001 so that the grid reynolds is sufficiently low.

Probably then move onto MITgcm...