Ran open ice shelf cavity with no melting for 5 days
Ran simulation for the open ice shelf cavity for 5 days forced by
restoring salinity of 34.41 with initial salinity of 34.4 PSU.

dt = 600s K=1e-2m^2/s output time 12 hours

The temperature field remained at 1degC as expected i.e advecting
a temperature of 1deg C outside the domain into 1degC inside.

As the simulation progressed the cells moved with the density front.
By 5 days the cells were about 3/4 of the way to the grounding line.
The density front was only about half way down the domain.

So the theory that the cells are to do with currents setup where the density
is still constant in the vertical seems to hold up. I have run the simulation
again starting with the checkpoint at 5 days for 90 days though will hopefully
reach a steady state before then.

We are expecting that the overturning cells will eventually be forced out of the
domain once the restoring salinity has flushed all the way to the grounding line
This should give one large cell which eventually will have completely replaced
the water inside the domain so that the salinity is 34.41 everywhere. At this point
there should be no more forcing so the simulation should reach a steady state.