Run closed structured ice shelf simulation with Kh=1e-3m^2/s for 1hr
Run a closed structured ice shelf simulation with Kh = 1e-3 m^2/s
and Kv = 1e^-4m^2/s with a timestep of 600 seconds. The melt param
also used gamma function of u fric based on the constants given
in Jenkins et al 2010.

On 8 cores this took a few minutes to run 1 hour of simulation.

The density solution seems stable. There is a very thin low density
diffusing out from the top boundary which is the fresh cold water
generated by melting. There are no cells/stripes forming in the
density fields which happened before the dg advection fixes.

The temperature and salinity fields do have
discontinuities between cells. This means the meltrate oscillates
rapidly along the ice shelf base. This is seen in Ben's FRISP results.

The velocity field also seems stable. The flow is clockwise with very
thin positive x component velocity along the ice shelf base. The y
velocities are also positive along the ice shelf base but they are
columns of alterning positive and negative velocity order 1e-6m/s.
Not sure if these will disappear as the simulation progresses.
The overturning cells don't seem to be the same as before the
dg advection fixes.

Prior to this I ran with Srestore 34.41 so that there was still a
0.01 PSU difference between the initial conditions and the forcing.
This still ran but in a closed box it shouldn't be needed because there
water can't flow in or out. The melting density structure was just superimposed
on the initial vertical density front seen in the no melting simulations.

-----------------------------------------------------------

I had also been running a longer version of the 5day simulation in
an open box that is forced by higher salinity values outside the domain.
I haven't looked at these yet but the simulation got to 17.5 days
before I accidentally crashed it by renaming the output folder.

Stephan found that the funny density stripes in the topright corner
of the no melting simulation is probably due to the slope limiters
preventing a baroclinic instability from forming. Need to talk with
Stephan to work out if this is a bad thing.

Matt emphasised that I need to compare early spinup with the spinup
of MITgcm. If you don't look at the first timesteps then possibly
you are missing important and sometimes weird results generated
by the model.

To do:

- Look at the long no melting simulation. Have the cells disappeared?
Has the density front reached the grounding line?

- Send Matt and Stephan the new melting simulation pictures

- Run melt simulation for longer to compare with MITgcm.

but....

I've realised there is a bit of a problem because the melt param
in MITgcm for the gamma(u*) is actually from the Holland and Jenkins 1999
paper which defines gammaT = u* / (Gamma_turb + Gamma_mole)

where the Gamma_turb also depends on u*.

I've tracked down where the maths in holland and jenkins 1999 appears
in MITgcm. The code is all in shelfice_thermodynamics and the constants
are mostly in shelfice_readparms, though some are calculated in
shelfice_thermodynamics. This is written in my notebook and will sort out
a more permanent notes tomorrow.

So for the comparison with MITgcm there are three options:

1) keep current our gamma(u* ) based on Jenkins et al 2010 and then change
line 0453 and 0455 in shelfice_thermodynamics to something like:
gammaT = GammaT u* where GammaT is the constant in Jenkins et al 2010.
Is this all I would have to do? This seems very risky.

2) write equations for holland and jenkins 1999 u* parameterisation so
that I can directly compare with the MITgcm friction velocity run which
presumably has been tested thoroughly... This should be less risky than tweaking
MITgcm but still could make mistakes.

3) run melt with constant gammaT and compare with constant gammaT MITgcm
since these are already set to the same values in both codes. This seems
like the least risky as I have already run MITgcm in this setting for the
high K=2m^2/s case.

- So this evening will set off constant gammaT run for 12 hours outputing
every timestep 600s.

- run constant gammaT for 300 days output every 12 hours. (hopefully it will
reach a steady state within about 100 days...) if tomorrow is running ok
maybe restart from a checkpoint and decrease output frequency.

- tomorrow run necessary mitgcm simulations. What should i do about
timestep? should i try MITgcm with a timestep of 600s... or leave
at 30s specified by Ben?

- tomorrow start thinking about writing the other code for fric melt param

- if all is going well add in coriolis

- turn off slope limiters and investigate baroclinic instability.
Do we also see this in the melt runs?

- work out what the averaging for u* and T,s fields is. pretty sure
they refer to horizontal averaging in the comments...