Use constant gammaT for ice_shelf_cavity_structured_z_closed.py
Changed melt param to use the constant gammaT parameterisation
based on the constants in Holland and Jenkins 1999. This is to
allow easy comparison with MITgcm before we write the u* dependence
for gammaT in Holland and Jenkins 1999 which is of the form:

gammaT = u* / ( GammaT_turb(u*) + GammaT_mole )    (Eq 14)

I am going to add another script which has the Jenkins et al 2010
gammaT = GammaT u* parameterisation so that I can run both at once.
This form effectively uses constant GammaT_turb.