Removed dirichlet conditions for T and S on rhs for closed box
Specifying dirichlet conditions for T and S for the closed box
causes stability issues for the no melting simulation which stephan
has been running. Removing the dirichlet conditions gets rid of the
funny wave train.

The diffusion term enforces the dirichlet condition along the entire
which isn't needed if there is restoring. If we had higher resolution
maybe it would be ok but it doesn't seem to be sstable as is.

I've removed the T and S dirichlet conditions on the rhs from the
two closed box ice shelf scripts.

I ran both these simulations overnight but both diverged.

The constant gammaT solution diverged first after 1 day and the
gammaT = GammaT u* solution diverged after 2.5 days. I ran with
quite a long time step (600s) compared to the MITgcm setup (30s).

Before Stephan pointed out the mistake for the dirichlet conditions
I was planning to rerun the simulations with a timestep of 30s.

Also important as noted in the previous commit we are not using the
same thermodynamics as MITgcm so we shouldn't necessarily expect the
same dynamics.

In light of this the early time results from the closed ice shelf runs
seem to be quite good, because the structure of the velocity fields was
very similar. MITgcm definitely had 'funny vertical stripes' in the
vertical velocities and large oscillations in the melt rate. The
melt rates were also the same order of magnitude as were the velocities.

For the first time we saw a concave melt profile with the gammaT =
GammaT u* run, which matched MITgcm and Ben's k-eps runs.
Before the melt rate profiles had always been convex since the
melt rate did not depend on the friction velocity.

Outstanding questions?

- Still haven't looked at the long run (17.5days) without melting in an open
box. What happens if MITgcm mimics this setup with an open boundary? Do you still
see overturning cellls being flushed out?

- What is a stable timestep for melting? Can Mitgcm run with a longer timestep?

- Need to run MITgcm for low Ks with constant gammaT.

- What happens if you turn off melting in MITgcm and just have restoring in
closed box. Does this look like Stephan's run without the wave train?

- How many extra features does MITgcm have that I have missed? Matt suggested
looking for velocity restoring...