Removed limiters from closed box ice shelf structured setup
I removed the limiters from the closed box setup since Stephan
had run the closed box no melting setup without limiters (and
dirichlet bcs for T and S on rhs) and this had run stably
without the funny wave train.

The previous commit I ran without T and S dirichlet conditons on the
rhs and both constant gammaT and ufric dependent run crashed.

constant gammaT diverged after 2.5 days (2x longer than before)

ufric diverged after 2 days (shorter than before!). This seems a
bit weird but not sure what is going on yet.

I had a quick look at both runs  and the structure of the density
and velocity field seemed pretty similar to before when T and S
dirichlet conditons were set.

Based on the two layer flow I calculated that the timestep based on
the buoyancy frequency with a density difference of 0.06 kg/m^3
(after 2 days) ought to be 30s.  Not sure if this is just a coincidence
that this is Ben's value, since our dz is 2x more resolved....

The plan was to run without limiters since Stephan's closed box
simulation worked with the same order of magnitude del rho, which
I took to mean that the buoyancy frequency might not be such a hard
constraint...

If this didn't make a difference (which based on the crashing at the
same time, implies it didnt) I was going to run MITgcm for dt 600s
to see if it diverges.

Then I was going to work out how to run dt =30 s efficiently. Presumably
it means using the iterative solver which needs to be tested. I imagine
if the direct and iterative solver gives the same answer up to 1hour
then in theory they should give the same answer in the future...
It might also be worth reducing the resolution in the domain.

As I was going back to MITgcm I noticed/remembered there are four more
differences (as well as the dt and melt paramaterisation differences!)

- reference temp and sal in the linear eos are 1degC and 34.4PSU in MITgcm
and 0degC and 35PSU in firedrake but betaT,S are the same in both so I am
pretty sure this shouldn't make a difference.
- MITgcm has no slip bc on the sides set to false whereas I've added no
slip on the rhs boundary weakly when closing the box.
- Restoring in MITgcm is not a linear profile. It is just a wall and
perhaps importantly it is actually only over the last 600m. I reckon
you can maybe see where the sponge region is in the 48hr vertical velocity
plot. Our restoring region is 1000m wide but the restoring frequency drops
linearly to zero over that region.

Plan now is to rerun with restoring region matching MITgcm i.e constant only
over the last 600m.