Freeslip on rhs boundary for ice shelf structured cavity setup
Removed no slip condition from ice shelf cavity structured closed
box setups to match MITgcm.

Ran with dt =600s and this blew up again after 1 day.

Ran with dt = 30 s and curiously this blew up before 1 day at 21
hours.

Ran with dt = 10s and this did get to 24 hours. I stopped this in
order to run MITgcm simulations. There is an output file every hour.

MITgcm with the exact same setup as before but now with gammaT =
constant and dt = 600s blows up after 13 hours!

Reducing the timestep back to 30s it is able to run to 2 days ok
but the velocities are much higher that with the gamma(ufric) run
and there is definitely an overturning cell moving left.

The melt also has the same shape as the firedrake run with higher
melts at the grounding line (apart from the last cell where it goes
to zero - bug?).

It seems like the peak velocities are 44 hours as by 48 hours the
velocities have dropped a little. Have run for 100 days to see
what happens.

These results seem to be in really good agreement with the dt=10s
firedrake run. The velocity structure is very similar and the temperature
field looks pretty good too. The melts aren't quite the same value
but the profile is the same shape and the same order of magnitude.

I've run MITgcm for 100 days, hopefully it will have reached a steady
state by then.

The problem I now have is that the firedrake constant gammaT run
with a timestep of 10s takes 30-40mins to run 1 hour worth of simulation.
​My plan is to try the iterative solver again and hope that this now
works after the dg advection fixes stephan made. ​I should be able to
compare the output with the first 24 hours from the direct solve.

Stephan has also suggested changing to a 50m by 1m mesh and then
run with dt = 30s. This way we should be matching MITgcm exactly.

1) Change the mesh to 50m by 1m. Run with dt =30s. Should be a
fairly significant speed up.

2) Add iterative solver script. will this work?

3) Add H&J1999 gamma(ufric) to compare with MITgcm gamma(ufric)
simulations.

4) Think about potential temperature conversion.