Updated 2.5d code
Updated code to run 2.5d ice shelf cavity.

- Include u component in |u| for calculating friction velocity used in
melt rate
- Include u component in |u| for vw drag and include vw components
in |u| for u component drag in momentum equation and 2.5d scalar u
equation.
- Added Iterative HJ99 script for HJ99 melt rate parameterisation
and with coriolis. These scripts have both been tested against MITgcm
at 20 days and 8 days respectively and are in very good agreement.
Timestep = 30s and Kh = 1e-3m^2/s. The coriolis run has since reached
50 days (the end of the simulation). Need to check.
- The main factor allowing these runs is that the iterative solver was
updated by Stephan in the previous commit to include an initial poisson
solve. This fixed the stability problems so that the model is able to
get past 2days where the direct solver kept diverging.
- meltrate tests script has numbers for potential temperature conversion.
At 1000db (~1000m) a temperature of 0degC converts to a potential temp of
-0.0448degC. We are modelling potential temperature (because it is nearly
a conservative quantity). Imagining that the model temp is -0.0448degC,
with the correction to in situ temperature the melt rate increases by 4%.
- Also tidied up melt rate parameterisation script