Added option to output dataframe for matplotlib plotting later
In 2d iterative script (most up to date) have added option
to output a dataframe containing output values in numpy arrays
ready for contour plotting in matplotlib. Currently outputing the
dataframe prevents outputing vtu files. Not sure this is really
a good idea. Should add more flexibility.

Added a semi finished script to generate pngs of mitgcm vs thwaites
runs side by side. Tested this for 12 hours for the first 12 outputs.
The idea is that the pngs can be stitched together to make a gif/video.
This could easily be adaptaed to make videos of different firedrake
spinups.

At the moment I am using imageio which works fairly well to stitch pngs
together. Not sure if this is the best though...

To do:
- At some point need to sanitise run script testing folder, maybe move a
finished script into a demo folder?

- tidy up demo folder. Pretty sure all scripts are broken.

- Rename plotting script (it evolved out of a normal runscript).

- Add melt rate and other thermodynamics plotting. Also integrated
melt and Gade T-S curve would be useful.