Added a first attempt at an ice shelf cavity runscript with k-epsilon
Important change which may be wrong is that I added an if statement
for the psi_flux expression so that it is only updated when a stress
is specified in the bc.

THis is a very rough attempt at an ice shelf cavity runscript with
rans model set to k-epsilon.

The runscripts is based on the iterative solver for the MITgcm comparison
with dx =50m dz=1m.

With Kh = 1e-3 and Kv =1e-6+GLS and a timestep of 300s the model crashes
after 10 days but has a qualitatively similar solution to Kh=1e-3
Kv = 1e-4 m^2/s run.

I also ran with Kh= Kv =1e-6m^2/s +GLS from a spinup after 50 days of constant
diffusivities this was only stable up to 1 day witha timestep of 60s. When
Kh = 1e-3 m^2/s it was stable up to 1 day with atimestep of 300s
 which motivated using a time step above.