Added 2d and 2.5d runscripts for ice step domain based on Kimura et al 2013
2d run seems to be stable with Kv= 0.1m^2/s near the open ocean and linearly
scaling down as height of the layers decreases towards the grounding line.
Run to 100 days ok.

Before this I tried the values of Kv = 1e-5m^2/s in the paper, but it crashed
before 1 day. There is some instability that seems to happen near the
grounding line wall when the domain has completely flushed out and
you get flow back into the cavity.

I tried constant Kv = 1e-4,1e-3 and 1e-2. 1e-4 and 1e-3 both crashed before 1 day
but 1e-2 ran to 44 days. by 20 days though an instability was forming near
the grounding line.

I tried a 2.5d setup because I thought (naively?) that the velocities would be
slower with the u component. This didnt seem to happen and with Kv = 0.1m^2/s
the model crashed at 21 days.

The aspect ratio of the grid dx/dz = 1e3/(100/9) = 90 at the grounding line
and is (1e3/1e2)= 10 at the ice front and outside the cavity. So from a numerics
point of view not so obvious that Kh should be much bigger than Kv. In the fluidity
paper Kh = 100m^2/s. I havent tried lower values of Kh in case that makes it better.

Also importantly I haven't tried lowering the timestep. (set at 200s)

Or the grid resolution. On the plus side the run with 4 cores is very fast
1min = 1 day. And the melt rate profile is smooth due to the layered mesh.
The magnitudes also seem about right 1e-5m/a cf with fluidity results. we
are using different constants for GammaT/S though. mainly out of laziness on my
part and not wanting to mess up the melt param last night.

I thought that layer convergence near the grounding line wall might be affecting
melt rates (which definitely go up and down with the incoming flow) so
I added the linear ramp back in. For Kv = 0.01 this made the 2d simulation
crash quicker only lasting at 21 days compared to 44.

At this point raised open ocean Kv by another order of magnitude to 0.1m^2/s
(THis is actually the value given in isomip for when convective instability's
occur on a 32 layer grid - here only 9 layers)

The run was stable up and reached 100 days. It was more diffuse than Kimura 13
though. i.e the bottom layer was 0.96 degc compared to 1.8 degc when the initial
conditons and source regions were 2degc. it is quite a slow restoring rate though
Trestoring = 10 days at edge and goes down to 30 days at 90km.

So reasons for why we can't exactly replicate kimura 13:
- I've made a silly mistake with the setup. could well be possible because I set it up
last night...
- There is something wrong with our setup/model so we are inherently less stable
(less numerical diffusion?) than Fluidity. If it is just numerical diffusion then
is this necessarily a bad thing?
- if this is the case would lowering timestep or increasing resolution help?
- possibly to do with ratio of Kv/Kh. which may be related to aspect  ratio of grid too.
- or something inherently unstable about the 2d setup... maybe in 3d more drag. should
be quite easy to check because I have the mesh. just need to update boundary ids. This
seems like wishful thinking though.

Also needed to update .geo file for layered mesh because without defining
a physical surface of the 2d mesh gmsh didnt save the 2d part of the mesh!
Also note to self firedrake needs boundary id tags not descriptions you can add
in gmsh.

Finally also updated the 2.5d mitgcm comparison run to include the option for
outputs of arrays needed for matplotlib plotting to compare with mitgcm.