unstructured z style mesh.
Currently it is 10km long with same triangular shape 2m wall at gl
and 100m at rhs.

10 layers on rhs and 2 on left. and x scaling has pretty much made
all of the dx spacing about 500m which was the aim with 21 nodes.

To ensure that it stays triangular x stretching has to be done before
y stretching. otherwise you get a parabola (?) profile on top.