Plot melt rate profile with correct sampling
Plotting melt rate profile at 40 days with the updated sampling.
Before I was sampling the melt rate every two grid cells. This is a bit of
a bodge though because I just loaded a checkpoint at 40 days and resampled the
melt so there isn't a continuous updated melt rate available for the whole
spin up. This commit is really a reminder about this.