Ran cavity on unstructured z mesh for MITgcm setup, seemed ok!
Created an unstructred z mesh with gmsh using isotropic 1m triangles
for an ice shelf cavity with a 2m high grounding line wall and 100m
ocean wall, but only 200m long so that you get 200 triangles in horizontal.
Then stretched byu 50x to get 10km long domain and anisotripic elements
roughly 1m high and 50m long.

Before this (which took a few minutes) I had been faffing about trying to get
triangles that were 0.5 left hand side of rectangle and 0.01m at rhs of rectangle.
I think if the rectangle was 25 long (and 1 high) then this produced 200 elements
in the horizontal direciton. I thought I could then stretch this rectangle
to get grounding zone ice cavity... But I didnt come up with a way to get all the
elements to have 50m in the horizontal. Anyway the above method for generating
the mesh worked fine and this is just a special case anyway. Need to work out
how to meshing best...

The run seemed fine for the unstructred mesh. It got to 5 days ok with
Kh = 1e-3m^2/s and Kv = 1e-4m^2/s, dt = 120s. For closed boundary and
wall-like restoring. The solution seemed pretty similar to before but a bit
hard to tell in paraview. Have run with matplotlib outputting so can compare
with the original structured mesh. Also need to check runtime because I think
it was probably quite a bit slower.