Added coarse 2d setup in prep for 3d
Added coarse 2d setup to compare with coarse MITgcm. Have run
but not plotted yet. It is based on the setup used with the adjoint.

Have added a runscript for 3d (but no mesh yet) for a 3d wedge
without ice step. I have run the MITgcm equivalent for a closed
box with and without coriolis but not looked at the results yet.

If all this goes smoothly I think the plan is to add in an ice step
but based loosely on the isomip+ domain because if MITgcm is running
then this will be the test.