First 3d tests
I started running a 3d setup in late July 2020.
Initially I tried to run a coarse version of the 2d cavity in 3d.
The initial pressure solve was taking a long time so we (Stephan and I)
tried adapting the pressure projection parameters for the intial pressure
solve. This sped up the solve a lot, but was not stable.
I went back to a simple box setup with no forcing to check that nothing
happened. Stuff did happen!

Went back to using a direct solver for the velocity pressure system for the
do nothing box test case. This did nothing. (including passive T and
S tracers). Ran the same test case with the direct initial pressure
solve and iterative velocity-pressure solver (i.e. what we had been
using for 2d) and this also did nothing so all good.

So really I should have been more patient waiting for the initial solve.
The problem with direct solvers is that they scale badly when the dofs increase
which is especially true in 3d.

So then ran 3d box with salinity restoring imbalance similar to
run we had done in 2d. Broadly this behaved as expected with a
density front propagating into the domain along the y direction,
while the Temperature stayed the same as the forcing value matched the
initial conditions. There was a slight assymetry with flow faster along
one side of the box than the other. This seemed to be to do with the mesh alignment
 and is likely because we are running at coarse resolution.
(dx = dy = 500m, dz = 2m , Kh = 1e-2m^2/s, Kv = 2e-4m^2/s)

Running the setup with double the horizontal resolution led to broadly similar pattern
but the 'errors' at the sides of the box were aligned with the new grid.
Since we are using triangles for top surface or tets in 3d we should expect
to get non zero gradients in the across flow direction. I did compare
this with Mitgcm, which was perfectly symmetrical with no flow in the
x direction. But MITgcm is a structured grid where the
x and y components are separated so it is not a surprise that there
is a difference between the models at this coarse resolution.

I did try a restoring region where the relaxation time varied linearly
 in case this made a difference but the propagating density front
looked very similar.

Then started using iterative solver for T and S solves and the velocity
prediction solve. This sped up the timestepping considerably. So that the
bottleneck was definitely the initial pressure solve. We need a stable
iterative solver for the initial pressure solve. (Without the pressure
solve everything blows up).

We then added melting back into the box, which behaved as expected with cold,
fresh water diffusing into the domain from the top boundary. Then
went back to the coarse 3d wedge setup. The initial pressure solve took 30mins
and it took 3mins to run 1hour of simulation so 1:20.

The solution looks believable with the north-south(y) flow dominating.
X velocity should be zero but it has picked up to an order of magnitude
less than y velocity with the noise aligned with the mesh. vertical velocity
was an order of magnitude lower.

Added in coriolis and as expected flow is deflected to the west (pretending
we are in southern hemisphere). Currently this is just f plane assumption.
This is fine for isomip+ but maybe we want to have beta plane too. The
initial spinup seemed to have elevated north-south return flow on both sides
of the cavity, but after 10 days the main flow was along the western edge.

Comparing with MITgcm the solution is very similar but the velocities are
lower especially in the initial spinup. There may be problems with
how I am plotting MITgcm fields (especially now that the grid is so coarse).

---------------------------------------------------------------

So outstanding problems?

- speeding up initial pressure solve
- small differences cf 3d melt wedge with MITgcm
- pressure correction solve taking over 200 iterations. Stephan suggested using
hypre instead of gamg as pc.

----------------------------------------------------------------

Descritpion of scripts:
---------------------------------------------------------------
- run_scripts_testing/ice_shelf_cavity/3d_wedge/ice_shelf_cavity_structured_z_closed_HJ99gamma_ufric_iterative_coarse3d.py

   Broken runscript which uses vp solver params for initial pressure
   solve which doesnt work.
---------------------------------------------------------------
- run_scripts_testing/ice_shelf_cavity/3d_wedge/ice_shelf_cavity_structured_z_closed_HJ99gammafric_iterative_coarse3d_direct_initial.py

  Working 3d wedge without coriolis. Uses direct intial pressure solve and then iterative solves for T,S,velocity prediction and pressure
	projection
---------------------------------------------------------------
- run_scripts_testing/ice_shelf_cavity/3d_wedge/ice_shelf_cavity_structured_z_closed_HJ99gammafric_iterative_coarse3d_direct_initial_coriolis.py
   Working 3d wedge with coriolis. Uses direct initial pressure solve and then iterative solves for T,S,velocity prediction and pressure
   projection
---------------------------------------------------------------
- run_scripts_testing/ice_shelf_cavity/3d_wedge/ice_shelf_cavity_structured_z_closed_no_melt_direct_coarse3d.py

  3d box (not ice shelf cavity wedge!) without melting. Solved using direct solver, so quite slow. Used as a benchmark for
  iterative solves.

----------------------------------------------------------------
- run_scripts_testing/ice_shelf_cavity/3d_wedge/ice_shelf_cavity_structured_z_closed_no_melt_iterative_coarse3d.py

  BROKEN 3d box (not ice shelf cavity wedge!) without melting. Solved using iterative solver with iterative initial pressure solve so
  blows up.
----------------------------------------------------------------
- run_scripts_testing/ice_shelf_cavity/3d_wedge/ice_shelf_cavity_structured_z_closed_no_melt_iterative_coarse3d_direct_initial.py
  3d box (not ice shelf cavity wedge!) without melting. Solved using iterative solver, but with direct initial pressure solve.

--------------------------------------------------------------
- run_scripts_testing/ice_shelf_cavity/3d_wedge/ice_shelf_cavity_structured_z_closed_no_melt_iterative_coarse3d_iterative_initial.py

  BROKEN 3d box (not ice shelf cavity wedge!) without melting. Solved using iterative solver with iterative initial pressure solve so
  blows up