3d wedge with coriolis setup using initial iterative solve.
This script uses Stephan's recent changes to the initial pressure
solve, where the momentum terms for the velocity prediction step
are all treated explicitly.

This script has been run to 10 days and gives the same solution
to the same run but with a direct initial solve.

Timings just over 30mins to do 12 hour simulation time on a single core. so about
1hour per day.