Added runscript for 2d coarse isomip+ with extruded mesh
Currently this setup doesn't run as there is an issue with the
mesh. Probably there is a problem with how I definied the ice
step. But also maybe the grounding line as the length of
the left hand wall seems to be zero.

Maybe I am not implementing Stephan's trick to workaround
zero measure facet issue (fd issue #1858)...