Add 2d coarse isomip plus setup with adjoint
Currently this this is running with an obscure error on
my computer (ese-wiscott), which has an untested firedrake adjoint.
 THe error looks similar to the error fixed by the patch
for pyadjoint recommended by Stephan in the first
adjoint related commit.

Another possible source of error is trying to load previous
results from a checkpoint. I don't think I did this before
successfullym but unfortunately it is not obvious from my
notes from May 2020 whether I tried this and showed it
didnt work for an existing setup...

I then went back to try the original ice_shelf_coarse.py setup
and realised the adjoint on this now has an error. Possibly about
adding a domain to a constant...

Going back to previous firedrake installation on ese-piggt01-2
to see if that makes a difference...