Define objective function with an approx step
Using a conditional in the definition of the objective function
seems to be causing an error.

Using a conditional out of the objective function doesn't
seem to cause an error.

We need some way of only integrating the melt under the ice shelf
not along the whole 'top' boundary marked by the extruded mesh.
The workaround I have at the moment is making a function which
is an approximation to a backwards step function based on a
tanh profile.

The change from off to on occurs over a distance 2pi/k. The
offset x0 is the centre of the step function. When checking the
function the shelf length as calculated by the step function
approach should return x0.

For this setup with the wavelength set to 2 grid cells (8km)
the difference in integrated melt rate seems to be 5% cf with using
the conditional approach. This seems like quite a large
difference...