Restore to stable density profile
Previous commit I pushed was a first attempt at using an on-off
vertical viscosity switch based on a local density gradient.
i.e +ve gradient = unstable -> Kv = 1e-1m^2/s
and -ve gradient = stable -> Kv = 1e-3m^2/s

This made the Kv = 1e-3m^2/s simulation run for longer (57 days).
Previously the simulation had been crashing around 26 days as the
front hit the rhs wall.

The density gradient (and as a result the vertical viscosity/diffusivity)
field was very blotchy though. Stephan suggested projecting
density anomaly (not full density field as we have subtracted off
hydrostatic term in the momentum equation), as opposed to
doing a solve with IBP to move derivative onto continuous test P1
function.

I've tested the 'project first' method on a static density field
and seemed to give the same results as before. So next to try is
using this as part of kappa switch in a dynamic run.

However, alongside this I also decreased the timestep by half
(to 900s) using the solve approach and this ended up getting to 100
days. (This had been the end time for previous simulations which
seemed to be close to a quasi steady state...)

I then checked to see if decreasing the timestep made the original
simulation (without the kappa switch) stable.

Using the original restoring values (i.e. constant T and S values
and jump in restoring rate):
- dt = 1800s diverged 26 days (as above i.e when front hits rhs
wall)
- dt = 900s diverged 84 days
- dt = 600s got to 100 days but then diverged before 110 days!
The initial spinup of dt = 900s and dt = 600s was very similar
but after 60 days the dt = 600s simuation becomes much more dynamic,
with the oscillations obviously starting from the rhs restoring
region.

Then I tried with the restoring rate linearly ramping up over the
sponge region again (still with constant T and S restoring
values though).
- dt = 1800s as above diverged when front hits rhs wall
- dt = 900s got to 100 days
- dt = 600s diverged at 57 days!

Before dt = 600s diverged it looked very similar to dt = 900s.
I then calculated a timestep limited by buoyancy frequency based
on salinity gradient observed on rhs just beyond the source region.
This came out as 574s, so I tried dt = 450s and this simulation
got to 100 days ok! It also looked identical to dt = 900s simulation.

Then I switched to using a stable density gradient using the
'warm' values from ISOMIP+ Ocean0 setup for the initial conditions
and the restoring values. Perhaps unsurprisingly this made a big difference!

Simulations with Kv =1e-3 and dt = 1800s are now stable. The vertical velocity
at the rhs seems a bit lower (6e-3 c.f over 1e-2 m/s) and crucially seems to
stabilise before 100 days. The spinup propogation is quite different
due to the initial stratification profile. There is obviously
a shock as melting turns on which means that the initial velocity
is not as simple as the overturning circulation, associated with the
ice pump mechanism. However, after about 50 days the initial stratification
patterns seems to be gone and the velocity seems similar to before. There
is quite an obvious difference in the salinity and temperature profiles
in the region next to the restoring region, but it is not causing
instabilities.

I also tested lower timesteps (dt = 900s and dt = 450s) and both these simulations
are very similar to dt = 1800s upto 100 days. I also ran dt = 1800s
for another 500 days. This ran ok and the solution at 100 days
was effectively the same as at 600 days.

Finally I varied Kv with dt=1800s. Kv =1e-1 and Kv = 1e-2 basically both not
spun up by 100 days. Kv = 1e-1 horizontally stratified! And Kv = 1e-2
looks like it was getting closer to Kv=1e-3 just a bit slower and wider boundary layer…

Kv = 1e-4 similar width of boundary layer / velocity to Kv = 1e-3 but much
noisier. It had obvious ‘leaking’ from the ice ocean boundary that MITgcm
was showing (and when we used unstable DQ1Q2 elemenet pair…). It  crashes
at 37 days which was just as the front was about to run up the ice front.
Probably not a bad thing that the simulation diverges because we want
to know if the input to the simulation is 'rubbish'.

Final thing! I tried on off switch using the new Kv =1e-3 stratified setup,
so that I had a benchmark to compare with when testing the new 'project first'
method. It turned out that Kappav = 1e-1 all along top and bottom boundary
immediately after the first time step which seemed odd! This meant the density
gradient was positve all along the top and bottom boundary.

Looking again at the new setup (which is initially stratified in T and S)
I think it might be to do with imposing neumann homogenous boundary
conditions on the top and bottom. i.e grad T . n = 0 (and for S too)
on the top and bottom. So that imposes zero slope in the top and
bottom cells. Effectively the T and S values are constant in these cells
i.e like P0DG -> which is what you see in the first timestep when the
stratification pattern immediately changes.

I still don’t think that should generate a +ve density gradient
in the top and bottom cells though… really it should be a zero density gradient…
I am hoping (but not confident) that doing the
project method will sort it out. Possibly something happens during the
solve which messes it up?